third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger/dynamic-script-tag-expected.txt

CONSOLE MESSAGE: line 2: DYNAMIC document.write()
CONSOLE MESSAGE: line 2: DYNAMIC <script>
CONSOLE MESSAGE: line 2: DYNAMIC document.write()
CONSOLE MESSAGE: line 2: DYNAMIC <script>
Tests that inline scripts and document.write scripts get different uiSourceCodes with different URLs. Bug 87119


Running: testOpenDevToolsAfterLoad
Both dynamic scripts were loaded.
Number of non-inline scripts with document url as a sourceURL: 0.

Running: testOpenDevToolsThenReload
Reloading page.
The first script with document url as a sourceURL to be seen is inline script.
Page reloaded.
